feat: change double to float, normalize improved
- normalize() now call just one time length()

+++ src/vec3.h
@@ -7,23 +7,22 @@
 class Vec3
 {
     public:
-        double x,y,z;
+        float x,y,z;
         Vec3();
-        Vec3(double x,double y,double z);
+        Vec3(float x,float y,float z);
         Vec3 operator+(const Vec3& v)const;
         Vec3& operator+=(const Vec3& v);
         Vec3 operator-(const Vec3& v)const;
-        Vec3 operator*(const double t)const;//Multiply a vector by a scalar
-        friend Vec3 operator*(const double t,const Vec3& v);//Multiply a scalar by a vector
+        Vec3 operator*(const float t)const;//Multiply a vector by a scalar
+        friend Vec3 operator*(const float t,const Vec3& v);//Multiply a scalar by a vector
         Vec3 operator*(const Vec3& v)const;
-        Vec3 operator/(const double t)const;
-        double operator[](const int i)const;
+        Vec3 operator/(const float t)const;
+        float operator[](const int i)const;
         Vec3 operator-()const;//Take the opposite number
-        double dot(const Vec3& v)const;//Vector dot product
+        float dot(const Vec3& v)const;//Vector dot product
         Vec3 cross(const Vec3& v)const;//Vector cross product
-        double length()const;
+        float length()const;
         Vec3 normalize()const;
         friend std::ostream& operator<<(std::ostream& os,const Vec3& v);
         bool nearZero()const;//check whether a vector is close to zero
 };
-
